Stop tracking files that should be ignored (already in .gitignore)
The files are from the IntelliJ IDE (Android Studio's parent) and
are normally ignored successfully with the .gitignore Android Studio
automatically generates for new Flutter projects.

However, since '.idea' was a directory above where '.gitignore' was
located, '.idea' was not ignored and some of its contents became
tracked unintentionally at some point. That issue was fixed only
recently with commit 2062f14; as of that commit Android
Studio-related files should now be ignored. This commit evicts
these unintentionally tracked files out of the repo.