Change background color of DrawerHeader to 'secondaryVariant'
It appears the  ColorScheme member 'secondary' is also used for the edge
effect when scrolling. Thus, there existed no contrast between the edge
effect and the DrawerHeader's background color. Using 'secondaryVariant'
for the background color allows for the edge effect to be made out when
hitting the top of the Drawer's ListView.