Revert "Reformat profile1.dart"
This reverts commit 50d1a3848483ab9a09d81b7b164e6fbb3f0d3efe.

...and resolves the merge conflict upon trying to do so.

In hindsight, it was silly of me to have made that commit at all. If all
I wanted to do was to have a reformatted version of profile1.dart for my
own reference/use, I should have either put the reformatted file in a
stash or moved it outside of the repo's tree entirely. I should not have
made it part of the tree. This commit undoes that.

# Conflicts:
#	lib/profiles/profile1.dart

P.S. On an unrelated note, I would like to note somewhere in this repo's
history (it will be in this commit message in this case) that within a
few hours after ef28731 was made, I discovered that Sean and I are also
responsible for a few files that last line line endings. So, it looks
Tyler may be the only one of us who hasn't done this.