Prevent Android's back button from exiting Ready and Profile pages
(From what I recall, iOS has a system-wide gesture akin to Android's
system-wide back button, so that'll need to be tested at some point,
too.)

Logic for this fix was placed into the router; it seemed a fitting place
given that the fullscreen and rotation logic is also focused there.

Now let's see if/when this GitHub integration works... fixes #1.