Add TODO so we know to upgrade path_provider in the future
The app won't run (later it ended up being it wouldn't even build) in(for)
Android in my (integrated or not) development environment. After removing all
references to simple_share and upgrading path_provider to the latest version
(1.6.5), the app would run. I believe this is all due to bug or deficiency in
path_provider that has been fixed since version ^0.4.1. See TODO text for the
rest.